remove 120s validation

+++ .adonisjs/server/routes.d.ts
@@ -4,6 +4,77 @@ type ParamValue = string | number | bigint | boolean
 
 export type ScannedRoutes = {
   ALL: {
+    'admin.organization.overview_all': { paramsTuple?: []; params?: {} }
+    'admin.organization.activity_all': { paramsTuple?: []; params?: {} }
+    'admin.organization.stats_all': { paramsTuple?: []; params?: {} }
+    'admin.organization.index': { paramsTuple?: []; params?: {} }
+    'admin.organization.store': { paramsTuple?: []; params?: {} }
+    'admin.organization.show': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.organization.update': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.organization.users': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.organization.store_user': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.organization.roles': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.audit_log.by_organization': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.user.stats_all': { paramsTuple?: []; params?: {} }
+    'admin.user.index_all': { paramsTuple?: []; params?: {} }
+    'admin.user.store_all': { paramsTuple?: []; params?: {} }
+    'admin.user.show_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.user.update_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.user.suspend_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.user.activate_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.user.destroy_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.vehicle.stats_all': { paramsTuple?: []; params?: {} }
+    'admin.vehicle.index_all': { paramsTuple?: []; params?: {} }
+    'admin.vehicle.store_all': { paramsTuple?: []; params?: {} }
+    'admin.vehicle.show_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.vehicle.update_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.vehicle.destroy_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device.stats_all': { paramsTuple?: []; params?: {} }
+    'admin.device.index_all': { paramsTuple?: []; params?: {} }
+    'admin.device.store_all': { paramsTuple?: []; params?: {} }
+    'admin.device.show_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device.update_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device.destroy_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device.assign_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device.unassign_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device.live_telemetry_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device.live_history_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device.sync_flespi_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device.flespi_status_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device.flespi_logs_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device.flespi_telemetry_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.command.flespi_catalog_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.command.stats_all': { paramsTuple?: []; params?: {} }
+    'admin.command.index_all': { paramsTuple?: []; params?: {} }
+    'admin.command.catalog': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.command.history_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.command.results_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.vehicle_group.index_all': { paramsTuple?: []; params?: {} }
+    'admin.vehicle_group.store_all': { paramsTuple?: []; params?: {} }
+    'admin.vehicle_group.show': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.vehicle_group.update': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.vehicle_group.destroy': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.vehicle_group.members': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.vehicle_group.assign': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.vehicle_group.unassign': { paramsTuple: [ParamValue,ParamValue]; params: {'id': ParamValue,'memberId': ParamValue} }
+    'admin.vehicle_group.audit_logs_of_group': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device_group.index_all': { paramsTuple?: []; params?: {} }
+    'admin.device_group.store_all': { paramsTuple?: []; params?: {} }
+    'admin.device_group.show': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device_group.update': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device_group.destroy': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device_group.members': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device_group.assign': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device_group.unassign': { paramsTuple: [ParamValue,ParamValue]; params: {'id': ParamValue,'memberId': ParamValue} }
+    'admin.device_group.audit_logs_of_group': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.audit_log.index': { paramsTuple?: []; params?: {} }
+    'admin.audit_log.actions': { paramsTuple?: []; params?: {} }
+    'admin.command.sync_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.command.send_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.command.cancel_all': { paramsTuple: [ParamValue,ParamValue]; params: {'id': ParamValue,'commandId': ParamValue} }
+    'admin.command.execute_all': { paramsTuple: [ParamValue,ParamValue]; params: {'id': ParamValue,'command': ParamValue} }
+    'admin.commands.cutEngineForbidden': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.commands.restoreEngineForbidden': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
     'auth.login': { paramsTuple?: []; params?: {} }
     'auth.me': { paramsTuple?: []; params?: {} }
     'auth.logout': { paramsTuple?: []; params?: {} }
@@ -15,11 +86,6 @@ export type ScannedRoutes = {
     'user.update': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
     'user.suspend': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
     'user.destroy': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
-    'user.index_all': { paramsTuple?: []; params?: {} }
-    'user.destroy_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
-    'user.show_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
-    'user.update_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
-    'user.suspend_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
     'vehicle.index': { paramsTuple?: []; params?: {} }
     'vehicle.store': { paramsTuple?: []; params?: {} }
     'vehicle.show': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
@@ -111,12 +177,48 @@ export type ScannedRoutes = {
     'immobilization.restore': { paramsTuple?: []; params?: {} }
   }
   GET: {
+    'admin.organization.overview_all': { paramsTuple?: []; params?: {} }
+    'admin.organization.activity_all': { paramsTuple?: []; params?: {} }
+    'admin.organization.stats_all': { paramsTuple?: []; params?: {} }
+    'admin.organization.index': { paramsTuple?: []; params?: {} }
+    'admin.organization.show': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.organization.users': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.organization.roles': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.audit_log.by_organization': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.user.stats_all': { paramsTuple?: []; params?: {} }
+    'admin.user.index_all': { paramsTuple?: []; params?: {} }
+    'admin.user.show_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.vehicle.stats_all': { paramsTuple?: []; params?: {} }
+    'admin.vehicle.index_all': { paramsTuple?: []; params?: {} }
+    'admin.vehicle.show_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device.stats_all': { paramsTuple?: []; params?: {} }
+    'admin.device.index_all': { paramsTuple?: []; params?: {} }
+    'admin.device.show_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device.live_telemetry_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device.live_history_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device.flespi_status_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device.flespi_logs_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device.flespi_telemetry_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.command.flespi_catalog_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.command.stats_all': { paramsTuple?: []; params?: {} }
+    'admin.command.index_all': { paramsTuple?: []; params?: {} }
+    'admin.command.catalog': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.command.history_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.command.results_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.vehicle_group.index_all': { paramsTuple?: []; params?: {} }
+    'admin.vehicle_group.show': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.vehicle_group.members': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.vehicle_group.audit_logs_of_group': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device_group.index_all': { paramsTuple?: []; params?: {} }
+    'admin.device_group.show': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device_group.members': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device_group.audit_logs_of_group': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.audit_log.index': { paramsTuple?: []; params?: {} }
+    'admin.audit_log.actions': { paramsTuple?: []; params?: {} }
     'auth.me': { paramsTuple?: []; params?: {} }
     'user.roles': { paramsTuple?: []; params?: {} }
     'user.index': { paramsTuple?: []; params?: {} }
     'user.show': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
-    'user.index_all': { paramsTuple?: []; params?: {} }
-    'user.show_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
     'vehicle.index': { paramsTuple?: []; params?: {} }
     'vehicle.show': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
     'device.index': { paramsTuple?: []; params?: {} }
@@ -159,12 +261,48 @@ export type ScannedRoutes = {
     'command.results': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
   }
   HEAD: {
+    'admin.organization.overview_all': { paramsTuple?: []; params?: {} }
+    'admin.organization.activity_all': { paramsTuple?: []; params?: {} }
+    'admin.organization.stats_all': { paramsTuple?: []; params?: {} }
+    'admin.organization.index': { paramsTuple?: []; params?: {} }
+    'admin.organization.show': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.organization.users': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.organization.roles': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.audit_log.by_organization': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.user.stats_all': { paramsTuple?: []; params?: {} }
+    'admin.user.index_all': { paramsTuple?: []; params?: {} }
+    'admin.user.show_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.vehicle.stats_all': { paramsTuple?: []; params?: {} }
+    'admin.vehicle.index_all': { paramsTuple?: []; params?: {} }
+    'admin.vehicle.show_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device.stats_all': { paramsTuple?: []; params?: {} }
+    'admin.device.index_all': { paramsTuple?: []; params?: {} }
+    'admin.device.show_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device.live_telemetry_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device.live_history_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device.flespi_status_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device.flespi_logs_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device.flespi_telemetry_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.command.flespi_catalog_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.command.stats_all': { paramsTuple?: []; params?: {} }
+    'admin.command.index_all': { paramsTuple?: []; params?: {} }
+    'admin.command.catalog': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.command.history_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.command.results_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.vehicle_group.index_all': { paramsTuple?: []; params?: {} }
+    'admin.vehicle_group.show': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.vehicle_group.members': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.vehicle_group.audit_logs_of_group': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device_group.index_all': { paramsTuple?: []; params?: {} }
+    'admin.device_group.show': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device_group.members': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device_group.audit_logs_of_group': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.audit_log.index': { paramsTuple?: []; params?: {} }
+    'admin.audit_log.actions': { paramsTuple?: []; params?: {} }
     'auth.me': { paramsTuple?: []; params?: {} }
     'user.roles': { paramsTuple?: []; params?: {} }
     'user.index': { paramsTuple?: []; params?: {} }
     'user.show': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
-    'user.index_all': { paramsTuple?: []; params?: {} }
-    'user.show_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
     'vehicle.index': { paramsTuple?: []; params?: {} }
     'vehicle.show': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
     'device.index': { paramsTuple?: []; params?: {} }
@@ -207,12 +345,29 @@ export type ScannedRoutes = {
     'command.results': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
   }
   POST: {
+    'admin.organization.store': { paramsTuple?: []; params?: {} }
+    'admin.organization.store_user': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.user.store_all': { paramsTuple?: []; params?: {} }
+    'admin.user.suspend_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.user.activate_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.vehicle.store_all': { paramsTuple?: []; params?: {} }
+    'admin.device.store_all': { paramsTuple?: []; params?: {} }
+    'admin.device.assign_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device.sync_flespi_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.vehicle_group.store_all': { paramsTuple?: []; params?: {} }
+    'admin.vehicle_group.assign': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device_group.store_all': { paramsTuple?: []; params?: {} }
+    'admin.device_group.assign': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.command.sync_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.command.send_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.command.execute_all': { paramsTuple: [ParamValue,ParamValue]; params: {'id': ParamValue,'command': ParamValue} }
+    'admin.commands.cutEngineForbidden': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.commands.restoreEngineForbidden': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
     'auth.login': { paramsTuple?: []; params?: {} }
     'auth.logout': { paramsTuple?: []; params?: {} }
     'user.change_password': { paramsTuple?: []; params?: {} }
     'user.store': { paramsTuple?: []; params?: {} }
     'user.suspend': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
-    'user.suspend_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
     'vehicle.store': { paramsTuple?: []; params?: {} }
     'device.store': { paramsTuple?: []; params?: {} }
     'device.assign': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
@@ -249,8 +404,13 @@ export type ScannedRoutes = {
     'immobilization.restore': { paramsTuple?: []; params?: {} }
   }
   PATCH: {
+    'admin.organization.update': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.user.update_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.vehicle.update_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device.update_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.vehicle_group.update': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device_group.update': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
     'user.update': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
-    'user.update_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
     'vehicle.update': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
     'device.update': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
     'organization.update': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
@@ -259,8 +419,16 @@ export type ScannedRoutes = {
     'flespi_channel.update': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
   }
   DELETE: {
+    'admin.user.destroy_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.vehicle.destroy_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device.destroy_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device.unassign_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.vehicle_group.destroy': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.vehicle_group.unassign': { paramsTuple: [ParamValue,ParamValue]; params: {'id': ParamValue,'memberId': ParamValue} }
+    'admin.device_group.destroy': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
+    'admin.device_group.unassign': { paramsTuple: [ParamValue,ParamValue]; params: {'id': ParamValue,'memberId': ParamValue} }
+    'admin.command.cancel_all': { paramsTuple: [ParamValue,ParamValue]; params: {'id': ParamValue,'commandId': ParamValue} }
     'user.destroy': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
-    'user.destroy_all': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
     'vehicle.destroy': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
     'device.destroy': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
     'device.unassign': { paramsTuple: [ParamValue]; params: {'id': ParamValue} }
